(MoritzLangenstein) Refactored code structure of metis and athena
