+++ athena/src/shared_fragment_tree/pickle/packing.rs
